requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-dotenv==1.1.0
pydantic-settings==2.8.1
openai==1.75.0
langchain==0.3.23
langchain-community==0.3.21
langchain-text-splitters==0.3.8
langchain-chroma==0.2.2
chromadb==0.6.3
sentence-transformers==3.4.1
beautifulsoup4==4.13.3
pypdf==5.4.0
numpy==1.26.4
pytest==8.3.5
httpx==0.28.1
